[CI] build TEST_FILES for CI based on build manifest
Within the CI, we know the location of the build manifest of
the OpenSearch Dashboards that got scaled up in prep for the
execution of these tests within this repo.

We can use that knowledge to determine the set of tests that
are valid against the distribution that this repo is being
executed on.

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>