[Security] fix tests that failed on ARM64
For some reason, message returned in the security plugins tests
differ from architectures. On x64 it will return as valid JSON
but on ARM64 it will be an object. The closest attempt to keeping
the message check was by stringifying them but on ARM64 it would
add the double quote.

So it fails to parse we can just ensure the resp code is correct.

Also switch to visible to not exist seems like we define this across
the repo as well.

Issue Resolved:
n/a

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>