[ISM] show all rows for test
When all the suites run together. There might an issue
with cleanup and the expected amount of rows on this page.
Here we ensure the largest amount of rows are shown because
within the test run the expected data stream index might be
on the next row if there are 20 plus indices.

Issue resolved:
n/a

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>