Wait for tenancy call if security enabled and bump version
Bumping to version 2.0.0, fix linter problem.

Also, preventing the test runner from navigating to a new
page when security plugin is enabled. The reason why is
because depending on how long the test takes to run, there
is an API call to refresh the permissions that take a little
bit longer than what tests are expecting. So if the test runner
makes the call to switch tenants but navigates to a new page
before that call is finished another call might run into a
permissions error. This ensures that if that call is made
then the test runner waits before navigating to the destination.

Issue:
https://github.com/opensearch-project/opensearch-dashboards-functional-test/issues/231

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>