[Improvement] fix issue with tenancy and test finder
Incorrectly merged the verification where I check if option to wait
for switching of tenant worked. Switched it back to the expected
scenario if the test passed the command it will wait.

Also, the test finder for manifest was missing the manifest path so
added that while also ensuring the minify OSD tests weren't ran
when verifying the distribution because those will fail as expected.

Issue:
n/a

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>