[CI] Update test template to not use timeout
GitHub workflow was failing because starting of OpenSearch
and therefore OpenSearch Dashboards would exit with 124.

For example:
https://github.com/opensearch-project/opensearch-dashboards-functional-test/actions/runs/3237324644/jobs/5304250104

This is due to us using a timeout function and it outputs a non-zero code
therefore github actions treats it as a failure code and doesn't run the test.

This fixes to utilize cypress to run the startup commands and tests
within a single command.

Issue:
n/a

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>