Fix Anomaly Detection System Indices Deletion in Cypress Tests
This commit addresses an issue where the deletion of Anomaly Detection (AD) system indices was failing in Cypress tests due to lack of super admin privileges. Instead of attempting to delete the indices as super admin, which isn't feasible in the Cypress environment, we now employ an API-based approach to delete the content of these AD indices.

Testing Performed:
1. Executed all tests in a security-enabled cluster confirming the successful deletion of AD system indices content.

Signed-off-by: Kaituo Li <kaituo@amazon.com>