Buffer `waitForLoader` before checking for icon
Some async calls occur after navigating and sometimes navigating too quickly
prevents the calls from having correct data and therefore the homeIcon is in
a bad state.

This adds the ability to buffer before checking for the homeIcon.

Defaulted to 0 ms, but for `yarn cypress:run-with-security` configures
WAIT_FOR_LOADER_BUFFER_MS to be 500 ms

Issue resolved:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/5028

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>