[Index-mgmt 2.10] Renamed index to avoid collision for other tests (#873)
During one of the tests for indices, we create an index with name reindex_opensearch_dashboards_sample_data_ecommerce that is superstring for the sample index opensearch_dashboards_sample_data_ecommerce. This new index is supposed to get deleted but sometimes it might not get deleted due to failure in backend and after that if some other test looks for the sample index and tries to select that index from a combo box, it faces collision and might end up not selecting any index. This PR renames the new index to avoid this situation.

### Issues Resolved

Fixes flaky test in Index Management Dashboards - opensearch-project/index-management-dashboards-plugin#880

Signed-off-by: Amardeepsingh Siglani <amardeep7194@gmail.com>