[Discover] set local storage directly for large string spec
Other test utilize the helper function and not seeing the particular need for this test
to ensure it switches. But by doing this avoid whatever is taking focus of the popover that
lets the test first click the button but before cypress test runner clicks the switch.

Issue:
n/a

Signed-off-by: Kawika Avilla <avillk@amazon.com>