Don't need to force DatePickerMenu open (#1358)
Due to https://github.com/opensearch-project/OpenSearch-Dashboards/pull/6167,
we don't have the double render issue when global state is updated.

Therefore we don't need to call makeDatePickerMenuOpen function, which is just
to reclick the DatePickerToggle to make the menu re-open. This function is a tmp
solution to allow test passing.

In this PR, we will remove the function and all the usages.

Signed-off-by: Anan <ananzh@amazon.com>
(cherry picked from commit 2881c0fc82fb7909ce41efa68dc6bf7f7f580ccc)